requirements.txt
fastapi
uvicorn[standard]
httpx
numpy
scikit-learn
sentence-transformers
google-generativeai
